refactor: organize whenever types into a dedicated subpackage
- Move InstantType and ZonedDateTimeType into the activemodel/types/whenever/ directory for better project structure.
- Export types in the new __init__.py file to simplify imports.

Generated-by: aiautocommit
